relocate state factory package
