+++ components/bsb/timer/state-builder/build.gradle
@@ -1,0 +1,21 @@
+plugins {
+    id("flipper.multiplatform")
+    id("flipper.multiplatform-dependencies")
+    id("flipper.anvil-multiplatform")
+    id("kotlinx-serialization")
+}
+
+commonDependencies {
+    implementation(projects.components.core.di)
+    implementation(projects.components.core.ktx)
+    implementation(projects.components.core.log)
+
+    implementation(projects.components.bsb.dao.api)
+    implementation(projects.components.bsb.timer.background.api)
+
+    implementation(libs.kotlin.datetime)
+}
+
+commonTestDependencies {
+    implementation(libs.kotlin.test)
+}
