+++ components/bsb/timer/state-builder/src/commonMain/kotlin/com/flipperdevices/bsb/timer/statefactory/TimerStateFactory.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.bsb.timer.background.statefactory
+package com.flipperdevices.bsb.timer.statefactory
 
 import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
 import com.flipperdevices.bsb.timer.background.model.TimerTimestamp
