+++ components/bsb/timer/state-builder/src/commonMain/kotlin/com/flipperdevices/bsb/timer/statefactory/iteration/model/IterationData.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.bsb.timer.background.api.statefactory.iteration.model
+package com.flipperdevices.bsb.timer.statefactory.iteration.model
 
 import kotlin.time.Duration
