+++ instances/bsb/build.gradle.kts
@@ -237,6 +237,7 @@ commonDependencies {
     implementation(projects.components.bsb.timer.rest.impl)
     implementation(projects.components.bsb.timer.cards.api)
     implementation(projects.components.bsb.timer.cards.impl)
+    implementation(projects.components.bsb.timer.stateBuilder)
 }
 
 dependencies {
