+++ settings.gradle.kts
@@ -154,6 +154,7 @@ include(
     ":components:bsb:timer:syncservice:firebase",
     ":components:bsb:timer:syncservice:noop",
     ":components:bsb:timer:syncservice:socket",
+    ":components:bsb:timer:state-builder",
 
     ":components:bsb:wear:bridge:messenger:api",
     ":components:bsb:wear:bridge:messenger:impl",
